Split `WebAssemblyToolchain` into `WASIToolchain` and `EmscriptenToolchain`
