--- Sources/SwiftDriver/Jobs/WebAssemblyToolchain+LinkerSupport.swift
@@ -1,287 +1,0 @@
-//===----------------------------------------------------------------------===//
-//
-// This source file is part of the Swift open source project
-//
-// Copyright (c) 2014 - 2025 Apple Inc. and the Swift project authors
-// Licensed under Apache License v2.0
-//
-// See LICENSE.txt for license information
-// See CONTRIBUTORS.txt for the list of Swift project authors
-//
-// SPDX-License-Identifier: Apache-2.0
-//
-//===----------------------------------------------------------------------===//
-
-import SwiftOptions
-
-import func TSCBasic.lookupExecutablePath
-import protocol TSCBasic.FileSystem
-import struct TSCBasic.AbsolutePath
-
-extension WebAssemblyToolchain {
-  public func addPlatformSpecificLinkerArgs(
-    to commandLine: inout [Job.ArgTemplate],
-    parsedOptions: inout ParsedOptions,
-    linkerOutputType: LinkOutputType,
-    inputs: [TypedVirtualPath],
-    outputFile: VirtualPath,
-    shouldUseInputFileList: Bool,
-    lto: LTOKind?,
-    sanitizers: Set<Sanitizer>,
-    targetInfo: FrontendTargetInfo
-  ) throws -> ResolvedTool {
-    let targetTriple = targetInfo.target.triple
-    switch linkerOutputType {
-    case .dynamicLibrary:
-      throw Error.dynamicLibrariesUnsupportedForTarget(targetTriple.triple)
-    case .executable:
-      let isEmscripten = targetTriple.os == .emscripten
-
-      // emcc already targets Emscripten, so skip --target and linker selection flags.
-      if !isEmscripten {
-        if !targetTriple.triple.isEmpty {
-          commandLine.appendFlag("-target")
-          commandLine.appendFlag(targetTriple.triple)
-        }
-
-        // Select the linker to use.
-        if let linkerArg = parsedOptions.getLastArgument(.useLd)?.asSingle {
-          commandLine.appendFlag("-fuse-ld=\(linkerArg)")
-        }
-
-        if let arg = parsedOptions.getLastArgument(.ldPath)?.asSingle {
-          commandLine.append(.joinedOptionAndPath("--ld-path=", try VirtualPath(path: arg)))
-        }
-      }
-
-      // Configure the toolchain.
-      //
-      // For Emscripten targets, use `emcc` (the Emscripten compiler driver) as the
-      // linker. emcc wraps wasm-ld and also generates JavaScript runtime glue.
-      //
-      // For other WebAssembly targets (WASI etc), use `clang` to perform the link.
-      // We use `clang` for the driver here because we do not wish to select a
-      // particular C++ runtime. Furthermore, until C++ interop is enabled, we
-      // cannot have a dependency on C++ code from pure Swift code. If linked
-      // libraries are C++ based, they should properly link C++. In the case of
-      // static linking, the user can explicitly specify the C++ runtime to link
-      // against. This is particularly important for platforms like android where as
-      // it is a Linux platform, the default C++ runtime is `libstdc++` which is
-      // unsupported on the target but as the builds are usually cross-compiled from
-      // Linux, libstdc++ is going to be present. This results in linking the wrong
-      // version of libstdc++ generating invalid binaries. It is also possible to
-      // use different C++ runtimes than the default C++ runtime for the platform
-      // (e.g. libc++ on Windows rather than msvcprt). When C++ interop is enabled,
-      // we will need to surface this via a driver flag. For now, opt for the simpler
-      // approach of just using `clang` and avoid a dependency on the C++ runtime.
-      var linkerPath: AbsolutePath
-      if isEmscripten {
-        linkerPath = try getToolPath(.emcc)
-        if let toolsDirPath = parsedOptions.getLastArgument(.toolsDirectory) {
-          let toolsDir = try AbsolutePath(validating: toolsDirPath.asSingle)
-          if let tool = lookupExecutablePath(filename: "emcc", searchPaths: [toolsDir]) {
-            linkerPath = tool
-          }
-          // emcc manages its own tool search paths and doesn't support
-          // the -B flag like clang does, so don't forward it.
-        }
-      } else {
-        linkerPath = try getToolPath(.clang)
-        if let toolsDirPath = parsedOptions.getLastArgument(.toolsDirectory) {
-          // FIXME: What if this isn't an absolute path?
-          let toolsDir = try AbsolutePath(validating: toolsDirPath.asSingle)
-
-          // If there is a clang in the toolchain folder, use that instead.
-          if let tool = lookupExecutablePath(filename: "clang", searchPaths: [toolsDir]) {
-            linkerPath = tool
-          }
-
-          // Look for binutils in the toolchain folder.
-          commandLine.appendFlag("-B")
-          commandLine.appendPath(toolsDir)
-        }
-      }
-
-      guard !parsedOptions.hasArgument(.noStaticStdlib, .noStaticExecutable) else {
-        throw Error.dynamicLibrariesUnsupportedForTarget(targetTriple.triple)
-      }
-
-      let runtimePaths = try runtimeLibraryPaths(
-        for: targetInfo,
-        parsedOptions: &parsedOptions,
-        sdkPath: targetInfo.sdkPath?.path,
-        isShared: false
-      )
-
-      if !parsedOptions.hasArgument(.nostartfiles) && !parsedOptions.isEmbeddedEnabled {
-        let swiftrtPath = VirtualPath.lookup(targetInfo.runtimeResourcePath.path)
-          .appending(
-            components: targetTriple.platformName() ?? "",
-            targetTriple.archName,
-            "swiftrt.o"
-          )
-        commandLine.appendPath(swiftrtPath)
-      }
-
-      let inputFiles: [Job.ArgTemplate] = inputs.compactMap { input in
-        // Autolink inputs are handled specially
-        if input.type == .autolink {
-          return .responseFilePath(input.file)
-        } else if input.type == .object {
-          return .path(input.file)
-        } else if lto != nil && input.type == .llvmBitcode {
-          return .path(input.file)
-        } else {
-          return nil
-        }
-      }
-      commandLine.append(contentsOf: inputFiles)
-
-      // emcc manages its own sysroot, so skip --sysroot for Emscripten.
-      if !isEmscripten {
-        // Prefer -sysroot as the native sysroot path if provided.
-        if let sysroot = parsedOptions.getLastArgument(.sysroot)?.asSingle {
-          commandLine.appendFlag("--sysroot")
-          try commandLine.appendPath(VirtualPath(path: sysroot))
-        } else if let path = targetInfo.sdkPath?.path {
-          commandLine.appendFlag("--sysroot")
-          commandLine.appendPath(VirtualPath.lookup(path))
-        }
-      }
-
-      // Add the runtime library link paths.
-      for path in runtimePaths {
-        commandLine.appendFlag(.L)
-        commandLine.appendPath(path)
-      }
-
-      let runtimeResourcePath = VirtualPath.lookup(targetInfo.runtimeResourcePath.path)
-      if parsedOptions.isEmbeddedEnabled {
-        // Allow linking certain standard library modules (`_Concurrency` etc)
-        let embeddedLibrariesPath: VirtualPath = runtimeResourcePath.appending(
-          components: "embedded", targetTriple.triple
-        )
-        if !isEmscripten {
-          commandLine.append(.flag("-Xlinker"))
-        }
-        commandLine.append(.joinedOptionAndPath("-L", embeddedLibrariesPath))
-      } else {
-        // Link the standard library and dependencies.
-        let linkFilePath: VirtualPath = runtimeResourcePath
-          .appending(components: targetTriple.platformName() ?? "",
-                     "static-executable-args.lnk")
-        guard try fileSystem.exists(linkFilePath) else {
-          throw Error.missingExternalDependency(linkFilePath.name)
-        }
-        commandLine.append(.responseFilePath(linkFilePath))
-      }
-
-      // Pass down an optimization level
-      if let optArg = mapOptimizationLevelToClangArg(from: &parsedOptions) {
-        commandLine.appendFlag(optArg)
-      }
-
-      // Explicitly pass the target to the linker (not needed for emcc)
-      if !isEmscripten {
-        commandLine.appendFlag("--target=\(targetTriple.triple)")
-      }
-
-      // WebAssembly doesn't reserve low addresses as its ABI. But without
-      // "extra inhabitants" of the pointer representation, runtime performance
-      // and memory footprint are significantly degraded. So we reserve the
-      // low addresses to use them as extra inhabitants by telling the lowest
-      // valid address to the linker.
-      // The value of lowest valid address, called "global base", must be always
-      // synchronized with `SWIFT_ABI_WASM32_LEAST_VALID_POINTER` defined in
-      // apple/swift's runtime library.
-      let SWIFT_ABI_WASM32_LEAST_VALID_POINTER = 4096
-      if isEmscripten {
-        // Use emcc settings instead of raw wasm-ld flags. Setting
-        // -sGLOBAL_BASE prevents emcc from auto-enabling --stack-first,
-        // which would place the stack at address 0 and violate Swift's
-        // assumption that addresses below LEAST_VALID_POINTER are unused.
-        commandLine.appendFlag("-sGLOBAL_BASE=\(SWIFT_ABI_WASM32_LEAST_VALID_POINTER)")
-        commandLine.appendFlag("-sTABLE_BASE=\(SWIFT_ABI_WASM32_LEAST_VALID_POINTER)")
-      } else {
-        commandLine.appendFlag(.Xlinker)
-        commandLine.appendFlag("--global-base=\(SWIFT_ABI_WASM32_LEAST_VALID_POINTER)")
-        commandLine.appendFlag(.Xlinker)
-        commandLine.appendFlag("--table-base=\(SWIFT_ABI_WASM32_LEAST_VALID_POINTER)")
-      }
-
-      // Set slightly higher than the default (64K) stack size so that basic
-      // workflows like Swift Testing can run within this limited stack space.
-      let SWIFT_WASM_DEFAULT_STACK_SIZE = 1024 * 128
-      if isEmscripten {
-        commandLine.appendFlag("-sSTACK_SIZE=\(SWIFT_WASM_DEFAULT_STACK_SIZE)")
-      } else {
-        commandLine.appendFlag(.Xlinker)
-        commandLine.appendFlag("-z")
-        commandLine.appendFlag(.Xlinker)
-        commandLine.appendFlag("stack-size=\(SWIFT_WASM_DEFAULT_STACK_SIZE)")
-      }
-
-      // Delegate to Clang for sanitizers. It will figure out the correct linker
-      // options.
-      if linkerOutputType == .executable && !sanitizers.isEmpty {
-        let sanitizerNames = sanitizers
-          .map { $0.rawValue }
-          .sorted() // Sort so we get a stable, testable order
-          .joined(separator: ",")
-        commandLine.appendFlag("-fsanitize=\(sanitizerNames)")
-      }
-
-      if parsedOptions.hasArgument(.profileGenerate) {
-        let libProfile = VirtualPath.lookup(targetInfo.runtimeResourcePath.path)
-          .appending(components: "clang", "lib", targetTriple.osName,
-                                 "libclang_rt.profile-\(targetTriple.archName).a")
-        commandLine.appendPath(libProfile)
-      }
-
-      if let lto = lto {
-        switch lto {
-        case .llvmFull:
-          commandLine.appendFlag("-flto=full")
-        case .llvmThin:
-          commandLine.appendFlag("-flto=thin")
-        }
-      }
-
-      // Run clang++ in verbose mode if "-v" is set
-      try commandLine.appendLast(.v, from: &parsedOptions)
-
-      // These custom arguments should be right before the object file at the
-      // end.
-      try commandLine.appendAllExcept(
-        includeList: [.linkerOption],
-        excludeList: [.l],
-        from: &parsedOptions
-      )
-      addLinkedLibArgs(to: &commandLine, parsedOptions: &parsedOptions)
-      if isEmscripten {
-        // emcc supports -Xlinker for passing flags to wasm-ld, but
-        // -Xclang-linker is clang-specific and must not be forwarded.
-        for linkerOpt in parsedOptions.arguments(for: .Xlinker) {
-          commandLine.appendFlag(.Xlinker)
-          commandLine.appendFlag(linkerOpt.argument.asSingle)
-        }
-        // -Xemcc-linker passes arguments directly to emcc.
-        try commandLine.appendAllArguments(.XemccLinker, from: &parsedOptions)
-      } else {
-        try addExtraClangLinkerArgs(to: &commandLine, parsedOptions: &parsedOptions)
-      }
-
-        // This should be the last option, for convenience in checking output.
-      commandLine.appendFlag(.o)
-      commandLine.appendPath(outputFile)
-      return try resolvedTool(.clang, pathOverride: linkerPath)
-    case .staticLibrary:
-      // We're using 'ar' as a linker
-      commandLine.appendFlag("crs")
-      commandLine.appendPath(outputFile)
-
-      commandLine.append(contentsOf: inputs.lazy.filter { $0.type != .autolink }.map { .path($0.file) })
-      return try resolvedTool(.staticLinker(lto))
-    }
-  }
-}
