+++ Tests/SwiftDriverTests/TripleTests.swift
@@ -1563,9 +1563,11 @@ import class TSCBasic.DiagnosticsEngine
     assertToolchain("x86_64-unknown-fuchsia", None.self)
     assertToolchain("x86_64-unknown-hermit", None.self)
     assertToolchain("wasm32-unknown-unknown", None.self)
-    assertToolchain("wasm32-unknown-wasi", WebAssemblyToolchain.self)
+    assertToolchain("wasm32-unknown-wasi", WASIToolchain.self)
     assertToolchain("wasm64-unknown-unknown", None.self)
-    assertToolchain("wasm64-unknown-wasi", WebAssemblyToolchain.self)
+    assertToolchain("wasm64-unknown-wasi", WASIToolchain.self)
+    assertToolchain("wasm32-unknown-emscripten", EmscriptenToolchain.self)
+    assertToolchain("wasm64-unknown-emscripten", EmscriptenToolchain.self)
     assertToolchain("avr-unknown-unknown", None.self)
     assertToolchain("avr", None.self)
     assertToolchain("lanai-unknown-unknown", None.self)
